Renamed navbar component
