+++ src/Hypzeh.Web/ClientApp/components/App/components/layout/index.jsx
@@ -1,4 +1,5 @@
 export { default as Main } from './Main';
-export { default as Navbar } from './Navbar';
 export { default as Panel } from './Panel';
 export { default as Page } from './Page';
+export { default as PageNavbar } from './PageNavbar';
+export { default as ProjectNavbar } from './ProjectNavbar';
